requirements.txt
feedparser==6.0.12
googlenewsdecoder==0.1.7
httpx==0.28.1
newspaper3k==0.2.8
openai==2.14.0
pandas==2.3.3
python-dotenv==1.2.1
python_pptx==1.0.2
Requests==2.32.5
urllib3==2.6.2
